cloud/src/movie/requirements.txt
# Optionally include additional dependencies here
redis
google-cloud
google-cloud-bigquery
requests
